Move EfxJsonReader to another namespace and folder
