+++ CMakeLists.txt
@@ -22,12 +22,13 @@ set(SOURCES
     src/snd_wav.cpp
     src/zone.cpp
     src/interface.cpp
-    src/snd_efx_reader.cpp
+    src/Config/EfxJsonReader.cpp
     src/Effects/GoldSrcOcclusionCalculator.cpp
     src/Effects/SteamAudioOcclusionCalculator.cpp
     src/Utilities/Fade.cpp
     src/Workarounds/NoWorkarounds.cpp
-    src/Workarounds/XFiWorkarounds.cpp)
+    src/Workarounds/XFiWorkarounds.cpp
+)
 
 add_library(MetaAudio SHARED ${SOURCES})
 target_compile_definitions(MetaAudio PRIVATE AUDIO_EXPORTS)
